Landing router (#1)
Get this project underway! I now have a place to put writing, and also
pull in an exemplify framework features and methods.

This sets up:
* The home, blog, and about pages
* A real simple theme. I'm thinking cyan and yellow.
* The first blog post

It's probably not optimal to put out a totally-not-SEO-optimized website
as a blog post but... I couldn't help using the framework when that's an
option. Gotta keep putting it through its paces.